Updated installation script.
Refactored and reworked the script to increase efficiency.

I've added some checks that decrease the workload of the script when running an update or if it fails during an installation due to whatever reason.

Added specific region downloading, unsure if this is any use but when checking the various version of the game I did notice some were labelled differently from others so depending on Region, it's probably best to download the correct one.

No longer requires Unity Hub to be installed. Instead it directly downloads the necessary unity files, moves across the ones required and dumps the rest to save space and installation time.

Reworked how updating works, just because a full rm -rf isn't necessary and instead the game works fine after updating without deleting every file, deeming it unnecessary to delete /Bin folder.